Correctly handle leaf directories as base path arguments
If a base path argument is already a leaf directory, leaf-directories returns it
unchanged. If the path is relative, it will then fail to compare with its own
absolute base path, causing the remapping to fail. Because the trailing
separator will normally be missing from the input in this case, even if an
absolute path is given, the leaf dir itself will not be part of the pathname
directory component so that only its parent directory will be replicated.

Both issues can be avoided by normalizing the base path arguments when searching
for leaf directories.